\name{addDataFrameToSQLiteDB}
\alias{addDataFrameToSQLiteDB}
\title{
Add Data Frame to DB 
}
\description{
Append a data frame to an existing SQLite table
}
\usage{
addDataFrameToSQLiteDB(con, tablename, df)
}
\arguments{
  \item{con}{
Connection to the db
}
  \item{tablename}{
Name of the table to add to
}
  \item{df}{
The data frame to add
}
}
\details{
Adds the data frame to the SQLite table
}
\value{
}
\references{
}
\author{
Jeffrey Wong
}
\note{
}

\seealso{
SQLiteQuery
}
\examples{
}
