% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_for_splits.R
\name{adjust_for_splits}
\alias{adjust_for_splits}
\title{Adjust For Splits}
\usage{
adjust_for_splits(df, symbol)
}
\arguments{
\item{df}{data.frame containing the prices/dividends to be adjusted}

\item{symbol}{Ticker of the subject company}
}
\description{
This function adjusts prices or dividends by past company splits
}
