\name{backTest}
\alias{backTest}
\title{
Back-testing Engine
}
\description{
A back-testing engine.  Users can write trading strategies and test their performance using historic data.  Currently, there is only support for testing with 1 asset.
}
\usage{
backTest(Symbols, historic, ndays = 50, strategy = .strategy.sample, funds = 1e+05)
}
\arguments{
  \item{Symbols}{
A vector of stock symbols.  Tests the performance of (strategy) on these symbols by pulling their historic data for the past (ndays) from Yahoo Finance.
}
  \item{historic}{
An optional argument.  RFinance can monitor stock prices and option prices itself, instead of pulling data from Yahoo Finance.  A user can pass in a data frame generated by RFinance's "monitor" function.  The trading strategy will be tested using this historic data.  
}
  \item{ndays}{
Amount of days to back test
}
  \item{strategy}{
A function that takes in a data frame 'historic', and amount of funds.  The strategy function should produce a list containing two vectors: action and quantity.  These vectors represent what trading action shoudl take place on each time interval, and how much should be traded.  See details for more.  
}
  \item{funds}{
How much money is available to trade with.
}
}
\details{
The strategy function should be of the form

function(historic, funds=100000) {
  (Algorithm to determine whether to buy, sell, or hold, and how much)
  The vectors action and quantity should correspond to the order of the historic data frame
  return (list(quantity=quantity, profits=profits))
}

This will allow RFinance to compute profits from trading.  
}
\value{
\item{historic }{The historic data that was downloaded}
\item{optimal }{A list of the form from OptimalTrades}
\item{mystrategy }{The decisions that were made by the strategy algorithm}
\item{percentofmax }{The percentage of the profits obtained by mystrategy relative to the optimal strategy}
}
\references{

}
\author{
Jeffrey Wong
}
\note{

}

\seealso{

}
\examples{
backTest("AAPL", ndays=5, funds=100000)
}

\keyword{ back-testing }
\keyword{ performance }
\keyword{ trading }
