% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_yahoo_url.R
\name{build_yahoo_url}
\alias{build_yahoo_url}
\title{Build Yahoo URL}
\usage{
build_yahoo_url(symbol, from, to, interval = "1d", event, handle)
}
\arguments{
\item{symbol}{Refers to the ticker of the company}

\item{from}{Initial date}

\item{to}{Final date}

\item{interval}{Specifies the time frame used in the data. Can be '1d', '1wk' or '1mo'}

\item{event}{What type of event is being recorded. Can be 'div' for dividends, 'history' for price history or 'split'.}

\item{handle}{handle object}
}
\description{
This function defines the standard Yahoo URL for the API call
}
