% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{calculate_alpha}
\alias{calculate_alpha}
\title{Provided a returns time series, it will return Jensen's Alpha}
\usage{
calculate_alpha(Ra, Rb, Rf = 0)
}
\arguments{
\item{Ra}{A time series of asset returns}

\item{Rb}{A time series of benchmark returns}

\item{Rf}{A time series of risk free rate}
}
\description{
Provided a returns time series, it will return Jensen's Alpha
}
