% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{calculate_sharpe_ratio}
\alias{calculate_sharpe_ratio}
\title{Provided a returns time series, it will return Sharpe Ratio}
\usage{
calculate_sharpe_ratio(R, Rf = 0)
}
\arguments{
\item{R}{A time series of asset returns}

\item{Rf}{A time series of risk free rate}
}
\description{
Provided a returns time series, it will return Sharpe Ratio
}
