% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_credentials.R
\name{check_credentials}
\alias{check_credentials}
\title{Check Credentials}
\usage{
check_credentials()
}
\description{
This function will check whether a user has logged in. In case it hasn't, it will default to the 'rfinance' user
}
