% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectionHandler.R
\docType{class}
\name{ConnectionHandler}
\alias{ConnectionHandler}
\title{ConnectionHandler class}
\description{
This class creates objects to handle connections to the database
}
\details{
This is a ConnectionHandler class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ConnectionHandler$new()}}
\item \href{#method-get_statements}{\code{ConnectionHandler$get_statements()}}
\item \href{#method-get_tickers}{\code{ConnectionHandler$get_tickers()}}
\item \href{#method-clone}{\code{ConnectionHandler$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionHandler$new(username = NULL, password = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_statements"></a>}}
\subsection{Method \code{get_statements()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionHandler$get_statements(ticker)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_tickers"></a>}}
\subsection{Method \code{get_tickers()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionHandler$get_tickers()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionHandler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
