% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_unix.R
\name{date_to_unix}
\alias{date_to_unix}
\title{Date To Unix}
\usage{
date_to_unix(Date)
}
\arguments{
\item{Date}{String that can be transformed into Date format.}
}
\description{
Basic 'util' function to transform date-like values into unix. This is necessary for the Yahoo API.
}
\keyword{internal}
