% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_statements.R
\name{get_statements}
\alias{get_statements}
\title{Get Statements}
\usage{
get_statements(ticker)
}
\arguments{
\item{ticker}{String with the ticker of the company.}
}
\value{
Returns a data.table with the financial statements
}
\description{
This function checks uses an existing connection handler to retrieve financial statements from the database.
}
\author{
Alejandro Jiménez Rico \email{aljrico@gmail.com}, \href{https://aljrico.com}{Personal Website}
}
