% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dividends.R
\name{get_dividends}
\alias{get_dividends}
\title{Dividend payments history}
\usage{
get_dividends(
  symbol,
  from = "1950-01-01",
  to = Sys.Date(),
  split_adjust = FALSE
)
}
\arguments{
\item{symbol}{Specifies the ticker or symbol the user wants to download the data of, in the form of a string. Example: 'MSFT'}

\item{from}{Minimum date to get data from}

\item{to}{Maximum date to get data of}

\item{split_adjust}{Boolean deciding if data should adjusted with splits}
}
\value{
Returns a tibble
}
\description{
This function downloads all available dividend payments history of a given company.
}
\examples{

dividends <- get_dividends('MSFT')
}
