% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prices.R
\name{get_prices}
\alias{get_prices}
\title{Functions to call and retrieve financial data from remote databases, by using company tickers (symbols). It uses publicly available API to download all data.}
\usage{
get_prices(symbol, from = "1970-01-01", to = "2030-01-01")
}
\arguments{
\item{symbol}{Specifies the ticker or symbol the user wants to download the data of, in the form of a string. Example: 'MSFT'}

\item{from}{Minimum date to get data from}

\item{to}{Maximum date to get data of}
}
\value{
Returns a tibble
}
\description{
Functions to call and retrieve financial data from remote databases, by using company tickers (symbols). It uses publicly available API to download all data.
}
\examples{

prices <- get_prices('MSFT')
}
\author{
Alejandro Jiménez Rico \email{aljrico@gmail.com}, \href{https://aljrico.github.io}{Personal Website}
}
