% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sp500.R
\name{get_sp500}
\alias{get_sp500}
\title{get_sp500}
\usage{
get_sp500(index = "sp500")
}
\arguments{
\item{index}{Indicates what index should the tickets be retrieved from. Currently only 'sp100', 'sp400' and 'sp500' are working.}
}
\value{
Returns a vector.
}
\description{
This function parses the most recent list of tickers for a given index
}
