% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_splits.R
\name{get_splits}
\alias{get_splits}
\title{Functions to call and retrieve financial data from remote databases, by using company tickers (symbols). It uses publicly available API to download all data.}
\usage{
get_splits(symbol, from = "1970-01-01", to = Sys.Date())
}
\arguments{
\item{symbol}{Specifies the ticker or symbol the user wants to download the data of, in the form of a string. Example: 'MSFT'}

\item{from}{Minimum date to get data from}

\item{to}{Maximum date to get data of}
}
\value{
Returns a tibble
}
\description{
Functions to call and retrieve financial data from remote databases, by using company tickers (symbols). It uses publicly available API to download all data.
}
\examples{

splits <- get_splits("MSFT")
}
