% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_statements.R
\name{get_income_statement}
\alias{get_income_statement}
\alias{get_balance_sheet_statement}
\alias{get_cash_flow_statement}
\title{Functions to call and retrieve financial data from remote databases, by using company tickers (symbols). It uses publicly available API to download all data.}
\usage{
get_income_statement(symbol)

get_balance_sheet_statement(symbol)

get_cash_flow_statement(symbol)
}
\arguments{
\item{symbol}{Specifies the ticker or symbol the user wants to download the data of, in the form of a string. Example: 'MSFT'}
}
\value{
Returns a data frame.
}
\description{
Functions to call and retrieve financial data from remote databases, by using company tickers (symbols). It uses publicly available API to download all data.
}
\examples{

df <- get_income_statement(symbol = "MSFT")
df <- get_balance_sheet_statement(symbol = "MSFT")
df <- get_cash_flow_statement(symbol = "MSFT")

}
\author{
Alejandro Jiménez Rico \email{aljrico@gmail.com}, \href{https://aljrico.github.io}{Personal Website}
}
