% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_symbols_list.R
\name{get_symbols_list}
\alias{get_symbols_list}
\title{Functions to call and retrieve financial data from remote databases, by using company tickers (symbols). It uses publicly available API to download all data.}
\usage{
get_symbols_list()
}
\value{
Returns an array.
}
\description{
Functions to call and retrieve financial data from remote databases, by using company tickers (symbols). It uses publicly available API to download all data.
}
\examples{

symbols <- get_symbols_list()
}
\author{
Alejandro Jiménez Rico \email{aljrico@gmail.com}, \href{https://aljrico.github.io}{Personal Website}
}
