% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tickers.R
\name{get_tickers}
\alias{get_tickers}
\title{Download list of tickers from a given public index}
\usage{
get_tickers(index = "sp500")
}
\arguments{
\item{index}{Indicates what index should the tickets be retrieved from. Currently only 'sp100', 'sp400' and 'sp500' are working.}
}
\value{
Returns a tibble.
}
\description{
This function parses the most recent list of tickers for a given index
}
\examples{

tickers_list <- get_tickers()
}
\author{
Alejandro Jiménez Rico \email{aljrico@gmail.com}, \href{https://aljrico.github.io}{Personal Website}
}
