% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_in_out.R
\name{log_out}
\alias{log_out}
\title{Log Out}
\usage{
log_out(environment = 1)
}
\arguments{
\item{environment}{Leave blank. Which environment should the connection object be placed on?}
}
\value{
Does not return anything
}
\description{
This function looks for the connection handler, removes it and closes all connections.
}
