\name{openSQLiteDB}
\alias{openSQLiteDB}
\title{
Open a SQLite Database File
}
\description{
Establish a connection to a sqlite database file
}
\usage{
openSQLiteDB(database)
}
\arguments{
  \item{database}{
Path to database
}
}
\details{
}
\value{
\item{m }{The dbDriver as in RSQLite}
\item{con }{The connection to the db}
}
\references{
}
\author{
Jeffrey Wong
}
\note{
}

\seealso{
RSQLite
}
\examples{
}
