\name{portfolio.getValue}
\alias{portfolio.getValue}
\title{
Portfolio Value
}
\description{
Calculates the current value of the portfolio
}
\usage{
portfolio.getValue(portfolio)
}
\arguments{
  \item{portfolio}{

The portfolio, usually kept as a text file and read in.
}
}
\details{

}
\value{
}
\references{

}
\author{

Jeffrey Wong
}
\note{

}

\seealso{

}
\examples{
data(portfolio)
tryCatch({
portfolio.getValue(portfolio)
})
}
