\name{portfolio.save}
\alias{portfolio.save}
\title{
Write a portfolio to a file
}
\description{
Writes the contents of a portfolio to a file
}
\usage{
portfolio.save(portfolio, filename = "portfolio")
}
\arguments{
  \item{portfolio}{

The portfolio, usually kept in a text file and read in.
}
  \item{filename}{

The filename to write to
}
}
\details{

}
\value{
}
\references{

}
\author{

Jeffrey Wong
}
\note{

}

\seealso{

}
\examples{
data(portfolio)
portfolio.save("portfolio1")
}
