\name{SQLiteQuery}
\alias{SQLiteQuery}
\title{
Query a SQL Database
}
\description{
Query a SQL Database
}
\usage{
SQLiteQuery(con, sql)
}
\arguments{
  \item{con}{
The connection to the database.  Can be opened using openSQLiteDB
}
  \item{sql}{
A character vector representing the desired sql query
}
}
\details{
Queries the database found in the db connection using the sql statement
}
\value{
Depends on the sql query
}
\references{
}
\author{
Jeffrey Wong
}
\note{
}

\seealso{
}
\examples{
}
