% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_company_profile.R
\name{get_company_profile}
\alias{get_company_profile}
\title{Basic Company Profile}
\usage{
get_company_profile(symbol)
}
\arguments{
\item{symbol}{Specifies the ticker or symbol the user wants to download the data of, in the form of a string.}
}
\value{
Returns a tibble
}
\description{
This function downloads the latest basic information of a given company. Full name, sector, last dividend, etc.
}
\examples{

company_profile <- get_company_profile('MSFT')
}
\author{
Alejandro Jiménez Rico \email{aljrico@gmail.com}, \href{https://aljrico.github.io}{Personal Website}
}
