% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dividends.R
\name{get_dividends}
\alias{get_dividends}
\title{Functions to call and retrieve financial data from remote databases, by using company tickers (symbols). It uses publicly available API to download all data.}
\usage{
get_dividends(
  symbol,
  from = "1970-01-01",
  to = Sys.Date(),
  split_adjust = TRUE
)
}
\arguments{
\item{symbol}{Specifies the ticker or symbol the user wants to download the data of, in the form of a string. Example: 'MSFT'}

\item{from}{Minimum date to get data from}

\item{to}{Maximum date to get data of}

\item{split_adjust}{Boolean deciding if data should adjusted with splits}
}
\value{
Returns a tibble
}
\description{
Functions to call and retrieve financial data from remote databases, by using company tickers (symbols). It uses publicly available API to download all data.
}
\examples{

dividends <- get_dividends('MSFT')
}
\author{
Alejandro Jiménez Rico \email{aljrico@gmail.com}, \href{https://aljrico.github.io}{Personal Website}
}
