% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_symbols_list.R
\name{get_symbols_list}
\alias{get_symbols_list}
\title{Get all available symbols}
\usage{
get_symbols_list(index = "any")
}
\arguments{
\item{index}{Includes only symbols that appear in determined indexes ('sp500'). Deafults to 'any'.}
}
\value{
Returns an array with all symbols available in the API.
}
\description{
Get all available symbols
}
\examples{

symbols <- get_symbols_list()
}
\author{
Alejandro Jiménez Rico \email{aljrico@gmail.com}, \href{https://aljrico.github.io}{Personal Website}
}
