% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tickers.R
\name{get_tickers}
\alias{get_tickers}
\title{Download list of tickers from a given public index}
\usage{
get_tickers()
}
\value{
Returns a vector.
}
\description{
This function returns the list of available tickers in the data base
}
\examples{

tickers <- rfinance::get_tickers()
}
