% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_in_out.R
\name{log_in}
\alias{log_in}
\title{Log In}
\usage{
log_in(username, password, environment = 1)
}
\arguments{
\item{username}{String containing the username.}

\item{password}{String containing the password.}

\item{environment}{Leave blank. Which environment should the connection object be placed on?}
}
\value{
Does not return anything
}
\description{
This function checks user credentials, opens a connection with the database and stores its handler in the global environment.
}
