% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_simulation.R
\docType{class}
\name{portfolio}
\alias{portfolio}
\title{portfolio class}
\description{
portfolio class description
}
\details{
This is a Portfolio class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-buy}{\code{portfolio$buy()}}
\item \href{#method-sell}{\code{portfolio$sell()}}
\item \href{#method-sell_everything}{\code{portfolio$sell_everything()}}
\item \href{#method-cash_in}{\code{portfolio$cash_in()}}
\item \href{#method-cash_out}{\code{portfolio$cash_out()}}
\item \href{#method-download_data}{\code{portfolio$download_data()}}
\item \href{#method-ownership_evolution}{\code{portfolio$ownership_evolution()}}
\item \href{#method-simulate}{\code{portfolio$simulate()}}
\item \href{#method-clone}{\code{portfolio$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-buy"></a>}}
\subsection{Method \code{buy()}}{
Spend some cash to buy an asset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{portfolio$buy(asset, date, number_assets = NULL, money = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{asset}}{Symbol specifying the asset to be bought.}

\item{\code{date}}{Date or date-like string that tells when should this asset be purchased.}

\item{\code{number_assets}}{How many shares of the asset should be exchanged. Defaults to NULL.}

\item{\code{money}}{How much money}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sell"></a>}}
\subsection{Method \code{sell()}}{
Sells a determined asset and gets cash back.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{portfolio$sell(asset, date, number_assets = NULL, money = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{asset}}{Symbol specifying the asset to be sold}

\item{\code{date}}{Date or date-like string that tells when should this asset be sold.}

\item{\code{number_assets}}{How many shares of the asset should be exchanged.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sell_everything"></a>}}
\subsection{Method \code{sell_everything()}}{
Sells all owned assets.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{portfolio$sell_everything(date)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{date}}{Date or date-like string that tells when should this asset be sold.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cash_in"></a>}}
\subsection{Method \code{cash_in()}}{
Add cash into the user portfolio.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{portfolio$cash_in(amount, date, report = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{amount}}{How much cash should be moved.}

\item{\code{date}}{Date or date-like string.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cash_out"></a>}}
\subsection{Method \code{cash_out()}}{
Withdraw cash from the user portfolio.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{portfolio$cash_out(amount, date, report = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{amount}}{How much cash should be moved.}

\item{\code{date}}{Date or date-like string.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-download_data"></a>}}
\subsection{Method \code{download_data()}}{
Downloads and stores all avilable historic prices of a given asset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{portfolio$download_data(asset)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{asset}}{Symbol specifying the asset.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ownership_evolution"></a>}}
\subsection{Method \code{ownership_evolution()}}{
Calculates the historical evolution of asset ownership
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{portfolio$ownership_evolution(date, orders = self$orders)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{date}}{Symbol specifying the asset.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-simulate"></a>}}
\subsection{Method \code{simulate()}}{
Compute portfolio simulation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{portfolio$simulate(until = max(self$orders$date))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{until}}{Date limit to which the simulation should be computed.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{portfolio$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
